[build-webkit-org] Update clang automatically for Safer-CPP-Checks https://bugs.webkit.org/show_bug.cgi?id=283091 rdar://139818398
Reviewed by Ryan Haddad.

Run the following steps in SaferCPPChecks on build-webkit-org to automatically build clang.

Add InstallCMake, InstallNinja, PrintClangVersion, CheckoutLLVMProject, and UpdateClang
to Shared/steps.py as a first step towards consolidating similar code in ews-build and build-webkit-org.

* Tools/CISupport/Shared/__init__.py: Added.

* Tools/CISupport/Shared/download-and-install-build-tools:
(download_and_install):
    - Add a command to remove the com.apple.quarantine attribute so that we don't hang on
      the first run of CMake.app.

* Tools/CISupport/Shared/steps.py:
    - Classes moved from ews-build. No new code changes were made.
(ShellMixin):
(AddToLogMixin):
(InstallCMake):
(InstallNinja):
(CheckOutLLVMProject):
(UpdateClang):
(PrintClangVersion):
(PrintClangVersionAfterUpdate):

* Tools/CISupport/Shared/steps_unittest.py:
    - Tests moved from ews-build. No new code changes were made.
(ExpectMasterShellCommand):
(BuildStepMixinAdditions):
(TestPrintClangVersion):
(TestCheckoutLLVMProject):
(TestUpdateClang):
(TestInstallCMake):
(TestInstallNinja):

* Tools/CISupport/build-webkit-org/factories.py:
(SaferCPPStaticAnalyzerFactory.__init__): Add steps mentioned above.

* Tools/CISupport/build-webkit-org/factories_unittest.py:
(TestExpectedBuildSteps): Add steps mentioned above.

* Tools/CISupport/ews-build/factories.py: Import Shared.

* Tools/CISupport/ews-build/steps.py: Deleted steps moved to Shared.

* Tools/CISupport/ews-build/steps_unittest.py: Deleted tests moved to Shared.
* Tools/CISupport/run-tests: Add 'Shared' to modules.

Canonical link: https://commits.webkit.org/286664@main
