+++ Tools/CISupport/ews-build/steps_unittest.py
@@ -43,24 +43,8 @@ from twisted.trial import unittest
 from . import send_email
 
 from .layout_test_failures import LayoutTestFailures
-from .steps import (AddReviewerToCommitMessage, AddMergeLabelsToPRs, AnalyzeAPITestsResults, AnalyzeCompileWebKitResults,
-                    AnalyzeJSCTestsResults, AnalyzeLayoutTestsResults, ApplyPatch, ApplyWatchList, ArchiveBuiltProduct, ArchiveTestResults, ArchiveStaticAnalyzerResults, BugzillaMixin, BlockPullRequest,
-                    Canonicalize, CheckOutPullRequest, CheckOutLLVMProject, CheckOutSource, CheckOutSpecificRevision, CheckChangeRelevance, CheckStatusOnEWSQueues, CheckStatusOfPR, CheckStyle,
-                    CleanBuild, CleanDerivedSources, CleanUpGitIndexLock, CleanGitRepo, CleanWorkingDirectory, CompileJSC, CompileJSCWithoutChange,
-                    CompileWebKit, CompileWebKitWithoutChange, ConfigureBuild, ConfigureBuild, Contributors, DetermineLabelOwner,
-                    DetermineLandedIdentifier, DisplaySaferCPPResults, DownloadBuiltProduct, DownloadBuiltProductFromMaster,
-                    EWS_BUILD_HOSTNAMES, ExtractBuiltProduct, ExtractTestResults,
-                    FetchBranches, FindModifiedLayoutTests, FindUnexpectedStaticAnalyzerResults, GetTestExpectationsBaseline, GetUpdatedTestExpectations, GitHub, GitHubMixin, GenerateS3URL,
-                    InstallGtkDependencies, InstallWpeDependencies, InstallHooks, LeaveComment, InstallCMake, InstallNinja,
-                    KillOldProcesses, ParseStaticAnalyzerResults, PrintConfiguration, PrintClangVersion, PushCommitToWebKitRepo, PushPullRequestBranch, RemoveAndAddLabels, ReRunAPITests, ReRunWebKitPerlTests, RetrievePRDataFromLabel,
-                    MapBranchAlias, ReRunWebKitTests, RevertAppliedChanges, RunAPITests, RunAPITestsWithoutChange, RunBindingsTests, RunBuildWebKitOrgUnitTests,
-                    RunBuildbotCheckConfigForBuildWebKit, RunBuildbotCheckConfigForEWS, RunEWSUnitTests, RunResultsdbpyTests,
-                    RunJavaScriptCoreTests, RunJSCTestsWithoutChange, RunWebKit1Tests, RunWebKitPerlTests,
-                    RunWebKitPyTests, RunWebKitTests, RunWebKitTestsInStressMode, RunWebKitTestsInStressGuardmallocMode,
-                    RunWebKitTestsWithoutChange, RunWebKitTestsRedTree, RunWebKitTestsRepeatFailuresRedTree,
-                    RunWebKitTestsRepeatFailuresWithoutChangeRedTree, RunWebKitTestsWithoutChangeRedTree, AnalyzeLayoutTestsResultsRedTree, TestWithFailureCount, SetBuildSummary, SetCommitQueueMinusFlagOnPatch,
-                    ScanBuild, ShowIdentifier, Trigger, TransferToS3, TwistedAdditions, UpdateClang, UpdatePullRequest, UpdateWorkingDirectory, UploadBuiltProduct,
-                    UploadFileToS3, UploadTestResults, ValidateCommitMessage, ValidateCommitterAndReviewer, ValidateChange, ValidateRemote, ValidateSquashed)
+from .steps import *
+from Shared.steps import *
 
 # Workaround for https://github.com/buildbot/buildbot/issues/4669
 from buildbot.test.fake.fakebuild import FakeBuild
@@ -9395,239 +9379,6 @@ class TestDisplaySaferCPPResults(BuildStepMixinAdditions, unittest.TestCase):
         self.assertEqual(self.getProperty('comment_text'), expected_comment)
         self.assertEqual(self.getProperty('build_finish_summary'), 'Found 10 new failures in File1.cpp')
         self.assertEqual([LeaveComment(), SetBuildSummary()], next_steps)
-
-
-class TestPrintClangVersion(BuildStepMixinAdditions, unittest.TestCase):
-    def setUp(self):
-        self.longMessage = True
-        return self.setUpBuildStep()
-
-    def tearDown(self):
-        return self.tearDownBuildStep()
-
-    def configureStep(self):
-        self.setupStep(PrintClangVersion())
-
-    def test_success(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir=LLVM_DIR,
-                        logEnviron=False,
-                        timeout=60,
-                        command=['./build/bin/clang', '--version'])
-            + ExpectShell.log('stdio', stdout='clang version 17.0.6 (https://github.com/rniwa/llvm-project.git 34715c1b2049d8aa738ade79f003ed4b82259a89) Target: arm64-apple-darwin23.5.0\nThread model: posix\nInstalledDir: /Volumes/Data/worker/macOS-Sonoma-Safer-CPP-Checks-EWS/llvm-project/./build/bin')
-            + 0,
-        )
-        self.expectOutcome(result=SUCCESS, state_string='clang version 17.0.6 (https://github.com/rniwa/llvm-project.git 34715c1b2049d8aa738ade79f003ed4b82259a89)')
-        rc = self.runStep()
-        self.assertEqual(self.getProperty('llvm_revision'), '34715c1b2049d8aa738ade79f003ed4b82259a89')
-        return rc
-
-    def test_failure(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir=LLVM_DIR,
-                        logEnviron=False,
-                        timeout=60,
-                        command=['./build/bin/clang', '--version'])
-            + ExpectShell.log('stdio', stdout='No such file or directory\n')
-            + 0,
-        )
-        self.expectOutcome(result=SUCCESS, state_string='Clang executable does not exist')
-        rc = self.runStep()
-        self.assertEqual(self.getProperty('llvm_revision'), None)
-        return rc
-
-
-class TestCheckoutLLVMProject(BuildStepMixinAdditions, unittest.TestCase):
-    LLVM_REVISION = '123456'
-
-    def setUp(self):
-        self.longMessage = True
-        return self.setUpBuildStep()
-
-    def tearDown(self):
-        return self.tearDownBuildStep()
-
-    def configureStep(self):
-        self.setupStep(CheckOutLLVMProject())
-
-        def doStepIf(self, step):
-            return self.build.getProperty('llvm_revision', '') != TestCheckoutLLVMProject.LLVM_REVISION
-        CheckOutLLVMProject.doStepIf = doStepIf
-
-    def test_skipped(self):
-        self.configureStep()
-        self.setProperty('llvm_revision', '123456')
-        self.expectOutcome(result=SKIPPED, state_string='llvm-project is already up to date')
-        return self.runStep()
-
-
-class TestUpdateClang(BuildStepMixinAdditions, unittest.TestCase):
-    ENV = {'PATH': '/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/Applications/CMake.app/Contents/bin/:BuildDir'}
-    LLVM_REVISION = '123456'
-
-    def setUp(self):
-        self.longMessage = True
-        return self.setUpBuildStep()
-
-    def tearDown(self):
-        return self.tearDownBuildStep()
-
-    def configureStep(self):
-        self.setupStep(UpdateClang())
-        self.setProperty('builddir', 'BuildDir')
-
-        def doStepIf(self, step):
-            return self.build.getProperty('llvm_revision', '') != TestUpdateClang.LLVM_REVISION
-        UpdateClang.doStepIf = doStepIf
-
-    def test_success(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir=LLVM_DIR,
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['/bin/sh', '-c', 'rm -r build-new; mkdir build-new']) + 0,
-            ExpectShell(workdir=LLVM_DIR,
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['/bin/sh', '-c', 'cd build-new; xcrun cmake -DLLVM_ENABLE_PROJECTS=clang -DCMAKE_BUILD_TYPE=Release -G Ninja ../llvm -DCMAKE_MAKE_PROGRAM=$(xcrun --sdk macosx --find ninja)']) + 0,
-            ExpectShell(workdir=LLVM_DIR,
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['/bin/sh', '-c', 'cd build-new; ninja clang']) + 0,
-            ExpectShell(workdir=LLVM_DIR,
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['rm', '-r', '../build/WebKitBuild']) + 0,
-        )
-        self.expectOutcome(result=SUCCESS, state_string='Successfully updated clang')
-        return self.runStep()
-
-    def test_skipped(self):
-        self.configureStep()
-        self.setProperty('llvm_revision', '123456')
-        self.expectOutcome(result=SKIPPED, state_string='Clang is already up to date')
-        self.runStep()
-
-
-class TestInstallCMake(BuildStepMixinAdditions, unittest.TestCase):
-    ENV = {'PATH': '/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/Applications/CMake.app/Contents/bin/'}
-
-    def setUp(self):
-        self.longMessage = True
-        return self.setUpBuildStep()
-
-    def tearDown(self):
-        return self.tearDownBuildStep()
-
-    def configureStep(self):
-        self.setupStep(InstallCMake())
-
-    def test_success_update(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir='wkdir',
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['python3', 'Tools/CISupport/Shared/download-and-install-build-tools', 'cmake'])
-            + ExpectShell.log('stdio', stdout='cmake version 3.30.4\n')
-            + 0,
-        )
-        self.expectOutcome(result=SUCCESS, state_string='Successfully installed CMake')
-        return self.runStep()
-
-    def test_success_update_skipped(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir='wkdir',
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['python3', 'Tools/CISupport/Shared/download-and-install-build-tools', 'cmake'])
-            + ExpectShell.log('stdio', stdout='cmake is already up to date... skipping download and installation.\n')
-            + 0,
-        )
-        self.expectOutcome(result=SUCCESS, state_string='CMake is already installed')
-        return self.runStep()
-
-    def test_failure(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir='wkdir',
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['python3', 'Tools/CISupport/Shared/download-and-install-build-tools', 'cmake'])
-            + ExpectShell.log('stdio', stdout='zsh: command not found: cmake\n')
-            + 1,
-        )
-        self.expectOutcome(result=FAILURE, state_string='Failed to install CMake')
-        return self.runStep()
-
-
-class TestInstallNinja(BuildStepMixinAdditions, unittest.TestCase):
-    ENV = {'PATH': "/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:BuildDir"}
-
-    def setUp(self):
-        self.longMessage = True
-        return self.setUpBuildStep()
-
-    def tearDown(self):
-        return self.tearDownBuildStep()
-
-    def configureStep(self):
-        self.setupStep(InstallNinja())
-        self.setProperty('builddir', 'BuildDir')
-
-    def test_success_update(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir='wkdir',
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['/bin/sh', '-c', 'cd ../; python3 build/Tools/CISupport/Shared/download-and-install-build-tools ninja'])
-            + ExpectShell.log('stdio', stdout='1.12.1\n')
-            + 0,
-        )
-        self.expectOutcome(result=SUCCESS, state_string='Successfully installed Ninja')
-        return self.runStep()
-
-    def test_success_update_skipped(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir='wkdir',
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['/bin/sh', '-c', 'cd ../; python3 build/Tools/CISupport/Shared/download-and-install-build-tools ninja'])
-            + ExpectShell.log('stdio', stdout='ninja is already up to date... skipping download and installation.\n')
-            + 0,
-        )
-        self.expectOutcome(result=SUCCESS, state_string='Ninja is already installed')
-        return self.runStep()
-
-    def test_failure(self):
-        self.configureStep()
-        self.expectRemoteCommands(
-            ExpectShell(workdir='wkdir',
-                        logEnviron=True,
-                        timeout=1200,
-                        env=self.ENV,
-                        command=['/bin/sh', '-c', 'cd ../; python3 build/Tools/CISupport/Shared/download-and-install-build-tools ninja'])
-            + ExpectShell.log('stdio', stdout='zsh: command not found: ninja')
-            + 1,
-        )
-        self.expectOutcome(result=FAILURE, state_string='Failed to install Ninja')
-        return self.runStep()
-
 
 if __name__ == '__main__':
     unittest.main()
